deprecate unused, respawn and batch

+++ server/dungeonsandtrolls/commands.go
@@ -64,7 +64,7 @@ func ValidateBuy(game *Game, identifiers *api.Identifiers, p *gameobject.Player)
 		if !ok {
 			return fmt.Errorf("%s is not an Item ID", id)
 		}
-		playerMoney -= i.BuyPrice
+		playerMoney -= int64(i.BuyPrice)
 		if playerMoney < 0 {
 			return fmt.Errorf("insufficient funds to make the purchase")
 		}
@@ -124,11 +124,11 @@ func ExecuteBuy(game *Game, p *gameobject.Player, identifiers *api.Identifiers) 
 		if !ok {
 			return fmt.Errorf("%s is not Item ID", itemId)
 		}
-		if p.Character.Money < item.BuyPrice {
+		if p.Character.Money < int64(item.BuyPrice) {
 			// this should not happen (thanks to the validation above)
 			return fmt.Errorf("insufficient funds to make the purchase")
 		}
-		p.Character.Money -= item.BuyPrice
+		p.Character.Money -= int64(item.BuyPrice)
 		buyEvent := api.Event_BUY
 		game.LogEvent(&api.Event{
 			Type: &buyEvent,
